CLM-31035: Remove UnnecessaryParentheses Rule (#83)
This is causing a lot of churn on insight-brain when upgrading. It seems to be catching more cases than it
used to in earliery versions of checkstyle. The team does not feel this rule is of value, 90% of the cases
it is catching are fealt by the team to be more readable as is